+++ inst/include/dplyr/DataFrameJoinVisitors.h
@@ -16,7 +16,13 @@ namespace dplyr{
             std::string name ;
             for( int i=0; i<nvisitors; i++){
                 name = names_[i] ;
-                visitors[i] = join_visitor( left[name], right[name], name) ;
+                try{
+                    visitors[i] = join_visitor( left[name], right[name], name) ;
+                } catch( ... ){
+                    std::stringstream s ;
+                    s << "cannot join on column '" << name << "'" ;
+                    stop(s.str());
+                }
             }
         }